+++ docs/examples/docs-demo-six-file-upload-5.vue
@@ -1,0 +1,16 @@
+<template>
+<div>
+
+        <six-file-upload uploading></six-file-upload>
+      
+</div>
+</template>
+<style>
+
+</style>
+<script>
+export default {
+  name: 'docs-demo-six-file-upload-5',
+  mounted() {  }
+}
+</script>