+++ libraries/ui-library/src/components/six-file-upload/test/six-file-upload.spec.tsx
@@ -17,15 +17,17 @@ describe('six-file-upload', () => {
           <div class="six-file-upload">
             <six-card class="six-file-upload__container--full">
               <div class="six-file-upload__drop-zone">
-                <span>
+                <div>
                   <span>
-                    Drop files to upload, or
-                    <span class="six-file-upload__label--highlighted">
-                      browse
+                    <span>
+                      Drop files to upload, or
+                      <span class="six-file-upload__label--highlighted">
+                        browse
+                      </span>
                     </span>
                   </span>
-                </span>
-                <input class="six-file-upload__input" name="resume" type="file">
+                  <input class="six-file-upload__input" name="resume" type="file">
+                </div>
               </div>
             </six-card>
           </div>
@@ -44,10 +46,12 @@ describe('six-file-upload', () => {
           <div class="six-file-upload">
             <six-card class="six-file-upload__container--full">
               <div class="six-file-upload__drop-zone">
-                <span>
-                  some custom label
-                </span>
-                <input class="six-file-upload__input" name="resume" type="file">
+                <div>
+                  <span>
+                    some custom label
+                  </span>
+                  <input class="six-file-upload__input" name="resume" type="file">
+                </div>
               </div>
             </six-card>
           </div>
@@ -64,17 +68,19 @@ describe('six-file-upload', () => {
     expect(page.root).toEqualHtml(`
         <six-file-upload disabled="">
           <div class="six-file-upload six-file-upload--disabled">
-            <six-card class="six-file-upload__container--full">
+            <six-card class="six-file-upload__container--full" disabled="">
               <div class="six-file-upload__drop-zone">
-                <span>
+                <div>
                   <span>
-                    Drop files to upload, or
-                    <span class="six-file-upload__label--highlighted">
-                      browse
+                    <span>
+                      Drop files to upload, or
+                      <span class="six-file-upload__label--highlighted">
+                        browse
+                      </span>
                     </span>
                   </span>
-                </span>
-                <input class="six-file-upload__input" disabled="" name="resume" type="file">
+                  <input class="six-file-upload__input" disabled="" name="resume" type="file">
+                </div>
               </div>
             </six-card>
           </div>
@@ -98,12 +104,65 @@ describe('six-file-upload', () => {
                 </six-icon>
               </span>
               <div class="six-file-upload__drop-zone six-file-upload__drop-zone--compact">
-                <span>
-                  Upload
-                </span>
-                <input class="six-file-upload__input" name="resume" type="file">
+                <div>
+                  <span>
+                    Upload
+                  </span>
+                  <input class="six-file-upload__input" name="resume" type="file">
+                </div>
               </div>
             </six-button>
+          </div>
+        </six-file-upload>
+    `);
+  });
+
+  it('renders compact and disabled', async () => {
+    const page = await newSpecPage({
+      components: [SixFileUpload],
+      html: `<six-file-upload compact disabled></six-file-upload>`,
+    });
+
+    expect(page.root).toEqualHtml(`
+        <six-file-upload compact="" disabled="">
+          <div class="six-file-upload six-file-upload--disabled">
+            <six-button class="six-file-upload__container--compact" disabled="">
+              <span slot="prefix">
+                <six-icon class="six-file-upload__label-icon">
+                  arrow_circle_up
+                </six-icon>
+              </span>
+              <div class="six-file-upload__drop-zone six-file-upload__drop-zone--compact">
+                <div>
+                  <span>
+                    Upload
+                  </span>
+                  <input class="six-file-upload__input" disabled="" name="resume" type="file">
+                </div>
+              </div>
+            </six-button>
+          </div>
+        </six-file-upload>
+    `);
+  });
+
+  it('renders uploading', async () => {
+    const page = await newSpecPage({
+      components: [SixFileUpload],
+      html: `<six-file-upload uploading></six-file-upload>`,
+    });
+
+    expect(page.root).toEqualHtml(`
+        <six-file-upload uploading="">
+          <div class="six-file-upload six-file-upload--disabled">
+            <six-card class="six-file-upload__container--full" disabled="">
+              <div class="six-file-upload__drop-zone">
+                <span class="six-file-upload__drop-zone__spinner-container">
+                  <six-spinner></six-spinner>
+                  Uploading...
+                </span>
+              </div>
+            </six-card>
           </div>
         </six-file-upload>
     `);